Moved utp.h to protocols directory
Moved utp.h to protocols directory and make necessary changes
to support this.
